[vm] move signature checking implementation to bytecode verifier
This is a more natural place to have these checks. Talked with @shaz about this.
