+++ language/bytecode_verifier/src/signature.rs
@@ -5,8 +5,14 @@
 //! parameters, locals, and fields of structs are well-formed. References can only occur at the
 //! top-level in all tokens.  Additionally, references cannot occur at all in field types.
 use vm::{
-    checks::SignatureCheck, errors::VerificationError, file_format::CompiledModule,
-    views::ModuleView, IndexKind,
+    access::ModuleAccess,
+    errors::{VMStaticViolation, VerificationError},
+    file_format::{CompiledModule, SignatureToken},
+    views::{
+        FieldDefinitionView, FunctionSignatureView, LocalsSignatureView, ModuleView,
+        TypeSignatureView, ViewInternals,
+    },
+    IndexKind, SignatureTokenKind,
 };
 
 pub struct SignatureChecker<'a> {
@@ -41,12 +47,11 @@ impl<'a> SignatureChecker<'a> {
             .fields()
             .enumerate()
             .filter_map(move |(idx, view)| {
-                view.check_signature_refs()
-                    .map(move |err| VerificationError {
-                        kind: IndexKind::FieldDefinition,
-                        idx,
-                        err,
-                    })
+                check_signature_refs(&view).map(move |err| VerificationError {
+                    kind: IndexKind::FieldDefinition,
+                    idx,
+                    err,
+                })
             })
             .collect();
         errors.push(signature_ref_errors);
@@ -69,4 +74,74 @@ impl<'a> SignatureChecker<'a> {
             .flatten()
             .collect()
     }
+}
+
+trait SignatureCheck {
+    fn check_signatures(&self) -> Vec<VMStaticViolation>;
+}
+
+impl<'a, T: ModuleAccess> SignatureCheck for FunctionSignatureView<'a, T> {
+    fn check_signatures(&self) -> Vec<VMStaticViolation> {
+        self.return_tokens()
+            .filter_map(|token| check_structure(token.as_inner()))
+            .chain(
+                self.arg_tokens()
+                    .filter_map(|token| check_structure(token.as_inner())),
+            )
+            .collect()
+    }
+}
+
+impl<'a, T: ModuleAccess> SignatureCheck for TypeSignatureView<'a, T> {
+    fn check_signatures(&self) -> Vec<VMStaticViolation> {
+        check_structure(self.token().as_inner())
+            .into_iter()
+            .collect()
+    }
+}
+
+impl<'a, T: ModuleAccess> SignatureCheck for LocalsSignatureView<'a, T> {
+    fn check_signatures(&self) -> Vec<VMStaticViolation> {
+        self.tokens()
+            .filter_map(|token| check_structure(token.as_inner()))
+            .collect()
+    }
+}
+
+/// Field definitions have additional constraints on signatures -- field signatures cannot be
+/// references or mutable references.
+pub(crate) fn check_signature_refs(
+    view: &FieldDefinitionView<'_, CompiledModule>,
+) -> Option<VMStaticViolation> {
+    let type_signature = view.type_signature();
+    let token = type_signature.token();
+    let kind = token.kind();
+    match kind {
+        SignatureTokenKind::Reference | SignatureTokenKind::MutableReference => Some(
+            VMStaticViolation::InvalidFieldDefReference(token.as_inner().clone(), kind),
+        ),
+        SignatureTokenKind::Value => None,
+    }
+}
+
+/// Check that this token is structurally correct.
+/// In particular, check that the token has a reference only at the top level.
+pub(crate) fn check_structure(token: &SignatureToken) -> Option<VMStaticViolation> {
+    use SignatureToken::*;
+
+    let inner_token_opt = match token {
+        Reference(token) => Some(token),
+        MutableReference(token) => Some(token),
+        Bool | U64 | String | ByteArray | Address | Struct(_) => None,
+    };
+    if let Some(inner_token) = inner_token_opt {
+        if inner_token.is_reference() {
+            return Some(VMStaticViolation::InvalidSignatureToken(
+                token.clone(),
+                token.kind(),
+                inner_token.kind(),
+            ));
+        }
+    }
+    None
 }
