+++ language/vm/src/checks.rs
@@ -2,7 +2,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod bounds;
-pub mod signature;
 
 pub use bounds::{BoundsCheck, BoundsChecker};
-pub use signature::SignatureCheck;
