+++ language/vm/src/unit_tests/mod.rs
@@ -1,7 +1,6 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-mod checks;
 mod deserializer_tests;
 mod fixture_tests;
 mod number_tests;
